requirement.txt
pillow==10.3
spotipy==2.24
moviepy==1.0.3 
opencv-python==4.10.0.84
google-api-python-client==2.147.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
replicate==0.34.1
google-api-python-client
yt-dlp